Refactor: split RemoteFeatureFlagStore to store and fetcher
