+++ WordPress/Classes/Stores/RemoteFeatureFlagStore.swift
@@ -1,5 +1,4 @@
 import Foundation
-import WordPressKit
 
 fileprivate extension DispatchQueue {
     static let remoteFeatureFlagStoreQueue = DispatchQueue(label: "remote-feature-flag-store-queue")
@@ -9,31 +8,13 @@ class RemoteFeatureFlagStore {
 
     /// Thread Safety Coordinator
     private var queue: DispatchQueue
-    private var persistenceStore: UserPersistentRepository
+    private var persistenceStore: UserDefaults
 
     init(queue: DispatchQueue = .remoteFeatureFlagStoreQueue,
-                 persistenceStore: UserPersistentRepository = UserDefaults.standard) {
+                 persistenceStore: UserDefaults = UserDefaults.standard) {
         self.queue = queue
         self.persistenceStore = persistenceStore
         DDLogInfo("🚩 Remote Feature Flag Device ID: \(deviceID)")
-    }
-
-    /// Fetches remote feature flags from the server.
-    /// - Parameter remote: An optional FeatureFlagRemote with a default WordPressComRestApi instance. Inject a FeatureFlagRemote with a different WordPressComRestApi instance
-    /// to authenticate with the Remote Feature Flags endpoint – this allows customizing flags server-side on a per-user basis.
-    /// - Parameter callback: An optional callback that can be used to update UI following the fetch. It is not called on the UI thread.
-    public func update(using remote: FeatureFlagRemote = FeatureFlagRemote(wordPressComRestApi: WordPressComRestApi.defaultApi()),
-                               then callback: FetchCallback? = nil) {
-        remote.getRemoteFeatureFlags(forDeviceId: deviceID) { [weak self] result in
-            switch result {
-                case .success(let flags):
-                    self?.cache = flags.dictionaryValue
-                    DDLogInfo("🚩 Successfully updated local feature flags: \(flags)")
-                    callback?()
-                case .failure(let error):
-                    DDLogError("🚩 Unable to update Feature Flag Store: \(error.localizedDescription)")
-            }
-        }
     }
 
     /// Checks if the local cache has a value for a given `FeatureFlag`
@@ -69,11 +50,9 @@ extension RemoteFeatureFlagStore {
         static let CachedFlagsKey = "FeatureFlagStoreCache"
     }
 
-    typealias FetchCallback = () -> Void
-
     /// The `deviceID` ensures we retain a stable set of Feature Flags between updates. If there are staged rollouts or other dynamic changes
     /// happening server-side we don't want out flags to change on each fetch, so we provide an anonymous ID to manage this.
-    private var deviceID: String {
+    var deviceID: String {
         guard let deviceID = persistenceStore.string(forKey: Constants.DeviceIdKey) else {
             DDLogInfo("🚩 Unable to find existing device ID – generating a new one")
             let newID = UUID().uuidString
@@ -85,7 +64,7 @@ extension RemoteFeatureFlagStore {
     }
 
     /// The local cache stores feature flags between runs so that the most recently fetched set are ready to go as soon as this object is instantiated.
-    private var cache: [String: Bool] {
+    var cache: [String: Bool] {
         get {
             // Read from the cache in a thread-safe way
             queue.sync {
