+++ WordPress/Classes/System/WordPressAppDelegate.swift
@@ -43,7 +43,7 @@ class WordPressAppDelegate: UIResponder, UIApplicationDelegate {
     private var pingHubManager: PingHubManager?
     private var noticePresenter: NoticePresenter?
     private var bgTask: UIBackgroundTaskIdentifier? = nil
-    private let remoteFeatureFlagStore = RemoteFeatureFlagStore()
+    private let remoteFeatureFlagStore = RemoteFeatureFlagsFetcher()
 
     private var mainContext: NSManagedObjectContext {
         return ContextManager.shared.mainContext
